# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/staff-tariff": {
            "post": {
                "description": "create a new staff tariff",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff-tariff"
                ],
                "summary": "Create a new staff tariff",
                "parameters": [
                    {
                        "description": "staff-Tariff",
                        "name": "staffTariff",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.CreateStaffTariff"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StaffTariff"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/staff-tariff/{id}": {
            "get": {
                "description": "get staff tariff by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff-tariff"
                ],
                "summary": "Get staff tariff by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "staff-tariff_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StaffTariff"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "get staff tariff",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff-tariff"
                ],
                "summary": "Update staff tariff",
                "parameters": [
                    {
                        "type": "string",
                        "description": "staff-tariff_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "staff-tariff",
                        "name": "staff-tariff",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateStaffTariff"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StaffTariff"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete staff tariff",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff-tariff"
                ],
                "summary": "Delete staff tariff",
                "parameters": [
                    {
                        "type": "string",
                        "description": "staff-tariff_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/staff-tariffs": {
            "get": {
                "description": "get staff tariff list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staff-tariff"
                ],
                "summary": "Get staff tariff list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StaffTariffResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateStaffTariff": {
            "type": "object",
            "properties": {
                "amount_for_card": {
                    "type": "integer"
                },
                "amount_for_cash": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tariff_type": {
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "description": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "models.StaffTariff": {
            "type": "object",
            "properties": {
                "amount_for_card": {
                    "type": "integer"
                },
                "amount_for_cash": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tariff_type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.StaffTariffResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "staff_tariffs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StaffTariff"
                    }
                }
            }
        },
...
        "models.UpdateStaffTariff": {
            "type": "object",
            "properties": {
                "amount_for_card": {
                    "type": "integer"
                },
                "amount_for_cash": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tariff_type": {
                    "type": "string"
                }
            }
        },
...
    }
}